requirements.txt
django-jazzmin
django-cleanup